backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
beautifulsoup4
selenium
PyPDF2
markdownify
python-dotenv
feedparser
lxml
webdriver-manager
tqdm
readability-lxml
pdfminer.six
PyMuPDF 